requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pillow==10.4.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
celery==5.4.0
redis==5.1.0
django-celery-beat==2.7.0
requests==2.32.3
django-cors-headers==4.4.0
drf-yasg==1.21.7
coverage==7.6.1
flake8==7.1.1
